Split out tmp_worktree into separate module for easy testing
